test(teams): split trigger integration scenarios [OPE-396]
